refactor: split artifact guardian into its own test file
validate_tasks_test.go had grown to ~1000 lines by mixing two unrelated
concerns: the validate-tasks command tests, and the package-wide artifact
guardian that reaps leaked test daemons and build dirs. The guardian is a
package-level lifecycle (TestMain + a re-exec'd reaper process), not
validate-tasks-specific, so hosting it in a command-named file obscured its
scope.

Move TestMain, the guardian machinery, and its regression/helper tests into
artifact_guardian_test.go, leaving validate_tasks_test.go for the command
tests, the shared binary builder, and workspace helpers. Package-level state
is split by owner: binary-builder vars stay with validateTasksBinary; guardian
vars move with the guardian. Pure relocation; no behavior change.
